requirements.txt
praw==7.7.1
prawcore>=2.0.0
huggingface_hub>=0.20.0
requests>=2.31.0
python-dotenv==1.0.0